backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
qdrant-client>=1.6.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-huggingface>=0.3.0
sentence-transformers>=2.2.0
tavily-python>=0.2.0
python-dotenv>=1.0.0
aiohttp>=3.8.5
dspy>=2.3.0
